requirements.txt
cryptography
PyQt6
gitpython
keyring